+++ src/main/java/com/kestrel/weddingbookkeeper/api/wedding/service/impl/QrService.java
@@ -5,7 +5,6 @@ import com.google.zxing.MultiFormatWriter;
 import com.google.zxing.WriterException;
 import com.google.zxing.client.j2se.MatrixToImageWriter;
 import com.google.zxing.common.BitMatrix;
-import com.kestrel.weddingbookkeeper.api.wedding.dto.request.WeddingInfoRequest;
 import com.kestrel.weddingbookkeeper.api.wedding.exception.QRCodeGenerationException;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -20,7 +19,7 @@ public class QrService {
 
     private static final String SERVICE_NAME = "weddingBookKeeper";
 
-    public InputStream generateQRCode(Integer weddingId) {
+    public InputStream generateQRCode(Long weddingId) {
         // QR 코드에 포함할 데이터 생성
         Map<String, Object> dataMap = new HashMap<>();
         dataMap.put("serviceName", SERVICE_NAME);
